requirements.txt
streamlit>=1.25.0
pandas
numpy
sentence-transformers
faiss-cpu
python-dotenv
cryptography
google-generativeai
